Extract loader component (and make it work)
