+++ app/components/Loader/index.js
@@ -1,0 +1,12 @@
+import { connect } from 'react-redux';
+import Loader from './presenter';
+
+function mapStateToProps(state) {
+  const sequence = state.sequence;
+
+  return {
+    display: sequence.loading,
+  };
+}
+
+export default connect(mapStateToProps)(Loader);
